Struct progress_tracker::Application

pub struct Application {
    pub widgets: Rc<Widgets>,
}

A Application is the foundation of an application. It wraps some low-level platform-specific services and is intended to act as the foundation for higher-level application classes such as GtkApplication or MxApplication. In general, you should not use this class outside of a higher level framework.

GApplication provides convenient life cycle management by maintaining a "use count" for the primary application instance. The use count can be changed using ApplicationExt::hold and ApplicationExt::release. If it drops to zero, the application exits. Higher-level classes such as GtkApplication employ the use count to ensure that the application stays alive as long as it has any opened windows.

Another feature that GApplication (optionally) provides is process uniqueness. Applications can make use of this functionality by providing a unique application ID. If given, only one application with this ID can be running at a time per session. The session concept is platform-dependent, but corresponds roughly to a graphical desktop login. When your application is launched again, its arguments are passed through platform communication to the already running program. The already running instance of the program is called the "primary instance"; for non-unique applications this is always the current instance. On Linux, the D-Bus session bus is used for communication.

The use of Application differs from some other commonly-used uniqueness libraries (such as libunique) in important ways. The application is not expected to manually register itself and check if it is the primary instance. Instead, the main function of a Application should do very little more than instantiating the application instance, possibly connecting signal handlers, then calling Application::run. All checks for uniqueness are done internally. If the application is the primary instance then the startup signal is emitted and the mainloop runs. If the application is not the primary instance then a signal is sent to the primary instance and Application::run promptly returns. See the code examples below.

If used, the expected form of an application identifier is the same as that of of a D-Bus well-known bus name. Examples include: com.example.MyApp, org.example.internal_apps.Calculator, org._7_zip.Archiver. For details on valid application identifiers, see Application::id_is_valid.

On Linux, the application identifier is claimed as a well-known bus name on the user's session bus. This means that the uniqueness of your application is scoped to the current session. It also means that your application may provide additional services (through registration of other object paths) at that bus name. The registration of these object paths should be done with the shared GDBus session bus. Note that due to the internal architecture of GDBus, method calls can be dispatched at any time (even if a main loop is not running). For this reason, you must ensure that any object paths that you wish to register are registered before Application attempts to acquire the bus name of your application (which happens in ApplicationExt::register). Unfortunately, this means that you cannot use ApplicationExt::get_is_remote to decide if you want to register object paths.

GApplication also implements the ActionGroup and ActionMap interfaces and lets you easily export actions by adding them with ActionMap::add_action. When invoking an action by calling ActionGroup::activate_action on the application, it is always invoked in the primary instance. The actions are also exported on the session bus, and GIO provides the DBusActionGroup wrapper to conveniently access them remotely. GIO provides a DBusMenuModel wrapper for remote access to exported GMenuModels.

There is a number of different entry points into a GApplication:

via 'Activate' (i.e. just starting the application)

via 'Open' (i.e. opening some files)

by handling a command-line

via activating an action

The Application::startup signal lets you handle the application initialization for all of these in a single place.

Regardless of which of these entry points is used to start the application, GApplication passes some ‘platform data’ from the launching instance to the primary instance, in the form of a glib::Variant dictionary mapping strings to variants. To use platform data, override the before_emit or after_emit virtual functions in your Application subclass. When dealing with ApplicationCommandLine objects, the platform data is directly available via ApplicationCommandLineExt::get_cwd, ApplicationCommandLineExt::get_environ and ApplicationCommandLineExt::get_platform_data.

As the name indicates, the platform data may vary depending on the operating system, but it always includes the current directory (key "cwd"), and optionally the environment (ie the set of environment variables and their values) of the calling process (key "environ"). The environment is only added to the platform data if the ApplicationFlags::SendEnvironment flag is set. Application subclasses can add their own platform data by overriding the add_platform_data virtual function. For instance, GtkApplication adds startup notification data in this way.

To parse commandline arguments you may handle the Application::command-line signal or override the local_command_line vfunc, to parse them in either the primary instance or the local instance, respectively.

For an example of opening files with a GApplication, see gapplication-example-open.c.

For an example of using actions with GApplication, see gapplication-example-actions.c.

For an example of using extra D-Bus hooks with GApplication, see gapplication-example-dbushooks.c.

Implements

ApplicationExt, glib::object::ObjectExt, ActionGroupExt, ActionMapExt, ApplicationExtManual Application is a class that handles many important aspects of a GTK+ application in a convenient fashion, without enforcing a one-size-fits-all application model.

Currently, Application handles GTK+ initialization, application uniqueness, session management, provides some basic scriptability and desktop shell integration by exporting actions and menus and manages a list of toplevel windows whose life-cycle is automatically tied to the life-cycle of your application.

While Application works fine with plain GtkWindows, it is recommended to use it together with ApplicationWindow.

When GDK threads are enabled, Application will acquire the GDK lock when invoking actions that arrive from other processes. The GDK lock is not touched for local action invocations. In order to have actions invoked in a predictable context it is therefore recommended that the GDK lock be held while invoking actions locally with gio::ActionGroup::activate_action. The same applies to actions associated with ApplicationWindow and to the “activate” and “open” gio::Application methods.

Automatic resources ## {automatic-resources}

Application will automatically load menus from the Builder resource located at "gtk/menus.ui", relative to the application's resource base path (see gio::ApplicationExt::set_resource_base_path). The menu with the ID "app-menu" is taken as the application's app menu and the menu with the ID "menubar" is taken as the application's menubar. Additional menus (most interesting submenus) can be named and accessed via GtkApplicationExt::get_menu_by_id which allows for dynamic population of a part of the menu structure.

If the resources "gtk/menus-appmenu.ui" or "gtk/menus-traditional.ui" are present then these files will be used in preference, depending on the value of GtkApplicationExt::prefers_app_menu. If the resource "gtk/menus-common.ui" is present it will be loaded as well. This is useful for storing items that are referenced from both "gtk/menus-appmenu.ui" and "gtk/menus-traditional.ui".

It is also possible to provide the menus manually using GtkApplicationExt::set_app_menu and GtkApplicationExt::set_menubar.

Application will also automatically setup an icon search path for the default icon theme by appending "icons" to the resource base path. This allows your application to easily store its icons as resources. See IconThemeExt::add_resource_path for more information.

If there is a resource located at "gtk/help-overlay.ui" which defines a ShortcutsWindow with ID "help_overlay" then Application associates an instance of this shortcuts window with each ApplicationWindow and sets up keyboard accelerators (Control-F1 and Control-?) to open it. To create a menu item that displays the shortcuts window, associate the item with the action win.show-help-overlay.

A simple application ## {gtkapplication}

A simple example

Application optionally registers with a session manager of the users session (if you set the Application:register-session property) and offers various functionality related to the session life-cycle.

An application can block various ways to end the session with the GtkApplicationExt::inhibit function. Typical use cases for this kind of inhibiting are long-running, uninterruptible operations, such as burning a CD or performing a disk backup. The session manager may not honor the inhibitor, but it can be expected to inform the user about the negative consequences of ending the session while inhibitors are present.

See Also ## {seealso}

HowDoI: Using Application, Getting Started with GTK+: Basics

Implements

GtkApplicationExt, gio::ApplicationExt, glib::object::ObjectExt, gio::ActionGroupExt, gio::ActionMapExt, gio::ApplicationExtManual

Fields
widgets: Rc<Widgets>
Implementations

impl Application

pub fn new(app: &Application) -> Self

Creates a new Application instance.

If non-None, the application id must be valid. See Application::id_is_valid.

If no application ID is given then some features of Application (most notably application uniqueness) will be disabled.

application_id

the application id

flags

the application flags

Returns

a new Application instance Creates a new Application instance.

When using Application, it is not necessary to call gtk_init manually. It is called as soon as the application gets registered as the primary instance.

Concretely, gtk_init is called in the default handler for the gio::Application::startup signal. Therefore, Application subclasses should chain up in their gio::Application::startup handler before using any GTK+ API.

Note that commandline arguments are not passed to gtk_init. All GTK+ functionality that is available via commandline arguments can also be achieved by setting suitable environment variables such as G_DEBUG, so this should not be a big problem. If you absolutely must support GTK+ commandline arguments, you can explicitly call gtk_init before creating the application instance.

If non-None, the application ID must be valid. See gio::Application::id_is_valid.

If no application ID is given then some features (most notably application uniqueness) will be disabled. A null application ID is only allowed with GTK+ 3.6 or later.

application_id

The application ID.

flags

the application flags

Returns

a new Application instance

pub(crate) fn connect_progress(&self)